See Description- Engine/ProtoGame2D
I have deleted the paddle class as this branch is meant for setting up the engine. I have also formatted the vertex data in meshbuilder to mimic the GL3 style. Also refactored some code. i have made a temporary test case to check if engine renderer is working and also incorporated the input system to it(this will be removed when i start working on the game)