See Description- Engine/Protogame2D
Finally started working on the Brick Breaker Game. I have created an Entity Base Class and its Child classes of the 3 components required for the game, the Paddle, the Ball and Brick. I have also completed temporary render of the paddle and the ball. Also tired to figure the collisions between the paddle and the ball, the paddle and the game bounds, the ball and the game bounds respectively. Now I am going to start working  on the bricks.